+++ SOURCE/server.cpp
@@ -90,12 +90,12 @@ void sendMessage_S(int socketFD, char* buffer)
 	//
 	// while (checkSend > 0);  // Loop forever until send buffer for this socket is empty
 
-	size_t len = sizeof(buffer);
+	int len = sizeof(buffer);
 	char *p = buffer;
-	ssize_t n;
+	int n;
 	while ( len > 0 && (n=send(socketFD,p,len,0)) > 0 ) {
 		p += n;
-		len =- (size_t)n;
+		len =- n;
 	}
 	if ( len > 0 || n < 0 ) {
 		// oops, something went wrong
@@ -125,12 +125,12 @@ void receiveMessage_S(int socketFD, char* buffer)
 	// 	exit(0);
 	// }
 
-	size_t len = sizeof(buffer);
+	int len = sizeof(buffer);
 	char *p = buffer;
-	ssize_t n;
+	int n;
 	while ( len > 0 && (n=recv(socketFD,p,len,0)) > 0 ) {
 		p += n;
-		len =- (size_t)n;
+		len =- n;
 	}
 	if ( len > 0 || n < 0 ) {
 		// oops, something went wrong
